+++ src/com/chess/engine/board/BoardUtils.java
@@ -1,5 +1,7 @@
 package com.chess.engine.board;
 
+import com.google.common.collect.ImmutableList;
+
 public class BoardUtils {
 
     public static final boolean[] FIRST_COLUMN = initColumn(0);
@@ -7,8 +9,8 @@ public class BoardUtils {
     public static final boolean[] SEVENTH_COLUMN = initColumn(6);
     public static final boolean[] EIGHT_COLUMN = initColumn(7);
 
-    public static final boolean[] SECOND_ROW = null;
-    public static final boolean[] SEVENTH_ROW = null;
+    public static final boolean[] SECOND_ROW = initRow(8);
+    public static final boolean[] SEVENTH_ROW = initRow(48);
 
     public static final int NUM_TILES = 64;
     public static final int NUM_TILES_PER_ROW = 8;
@@ -26,6 +28,17 @@ public class BoardUtils {
         } while (columnNumber < NUM_TILES);
 
         return column;
+    }
+
+    private static boolean[] initRow(int rowNumber) {
+        final boolean[] row = new boolean[NUM_TILES];
+
+        do {
+            row[rowNumber] = true;
+            rowNumber++;
+        } while (rowNumber % NUM_TILES_PER_ROW != 0);
+
+        return row;
     }
 
     /**