+++ src/com/chess/engine/board/Tile.java
@@ -40,6 +40,11 @@ public abstract class Tile {
         }
 
         @Override
+        public String toString() {
+            return "[]";
+        }
+
+        @Override
         public boolean isTileOccupied() {
             return false;
         }
@@ -56,6 +61,11 @@ public abstract class Tile {
         private OccupiedTile(int tileCoordinate, final Piece pieceOnTile) {
             super(tileCoordinate);
             this.pieceOnTile = pieceOnTile;
+        }
+
+        @Override
+        public String toString() {
+            return getPiece().getPieceAlliance().isBlack() ? getPiece().toString().toLowerCase() : getPiece().toString();
         }
 
         @Override