+++ src/com/chess/engine/pieces/Piece.java
@@ -18,6 +18,10 @@ public abstract class Piece {
         this.isFirstMove = false; // TODO: more work here!
     }
 
+    public int getPiecePosition() {
+        return this.piecePosition;
+    }
+
     public Alliance getPieceAlliance() {
         return this.pieceAlliance;
     }
@@ -28,4 +32,23 @@ public abstract class Piece {
 
     public abstract Collection<Move> calculateLegalMoves(final Board board);
 
+    public enum PieceType {
+        PAWN("P"),
+        KNIGHT("KN"),
+        BISHOP("B"),
+        ROOK("R"),
+        QUEEN("Q"),
+        KING("K");
+
+        private String pieceName;
+
+        PieceType(final String pieceName) {
+            this.pieceName = pieceName;
+        }
+
+        @Override
+        public String toString() {
+            return this.pieceName;
+        }
+    }
 }