+++ CHANGELOG.md
@@ -32,5 +32,6 @@
 - [Queen Class](https://en.wikipedia.org/wiki/Queen_(chess))
 - [Pawn Class](https://en.wikipedia.org/wiki/Pawn_(chess))
 - [King Class](https://en.wikipedia.org/wiki/King_(chess))
+- Board Class
 
 <!-- [0.1.0]: https://github.com/scriptjumper/Java-Chess-Engine/releases/tag/v-0.1.0 -->